+++ test/fixtures/scenarios/multisig_2_of_3.sql
@@ -5,6 +5,20 @@
 
 BEGIN TRANSACTION;
 
+-- Table  assets
+DROP TABLE IF EXISTS assets;
+CREATE TABLE assets(
+                      asset_id TEXT UNIQUE,
+                      asset_name TEXT UNIQUE,
+                      block_index INTEGER);
+INSERT INTO assets VALUES('0','BTC',NULL);
+INSERT INTO assets VALUES('1','XCP',NULL);
+INSERT INTO assets VALUES('18279','BBBB',310005);
+INSERT INTO assets VALUES('18280','BBBC',310006);
+-- Triggers and indices on  assets
+CREATE INDEX id_idx ON assets (asset_id);
+CREATE INDEX name_idx ON assets (asset_name);
+
 -- Table  balances
 DROP TABLE IF EXISTS balances;
 CREATE TABLE balances(
@@ -707,7 +721,6 @@ INSERT INTO order_matches VALUES('b898ed7f73ba34981adde222bf27498f66f352254817b0
 -- Triggers and indices on  order_matches
 CREATE INDEX backward_status_idx ON order_matches (backward_asset, status);
 CREATE INDEX forward_status_idx ON order_matches (forward_asset, status);
-CREATE INDEX id_idx ON order_matches (id);
 CREATE INDEX match_expire_idx ON order_matches (status, match_expire_index);
 CREATE INDEX tx0_address_idx ON order_matches (tx0_address);
 CREATE INDEX tx1_address_idx ON order_matches (tx1_address);