+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Client Versions ##
-* v9.49.0 (???)
+* v9.50.0 (???)
 	* ‘destroy’ functionality (protocol change: 335000)
 	* match IDs now include underscore
 	* code clean‐up
@@ -8,6 +8,10 @@
 	* handle protocol changes more elegantly
 	* more sophisticated version checking
 	* removed obsolete `carefulness` CLI option
+* v9.49.0 (2014-12-11)
+	* hotfix: numeric asset names (protocol change: 334000)
+	* sanity check on transaction construction
+	* minor bug fixes
 * v9.48.0 (2014-12-08)
 	* better enforcement of single source, destination
 * v9.47.1 (2014-12-02)